+++ packages/implementations/react/src/atomic/Input/src/components/Leading/index.ts
@@ -1,0 +1,4 @@
+import { Leading } from "./Leading";
+
+export { Leading } from "./Leading";
+export default Leading;