+++ packages/implementations/react/src/atomic/Input/src/components/Search/search.types.tsx
@@ -1,0 +1,10 @@
+import { InputHTMLAttributes } from "react";
+import { InputProperties } from "../../input.types";
+
+export interface Search {}
+
+export type SearchProperties = Search &
+  Pick<InputProperties, "divider" | "appearance" | "size">;
+
+export type _SearchProps = SearchProperties &
+  Omit<InputHTMLAttributes<HTMLInputElement>, "size">;